requirements.txt
PyYAML==6.0.1
fabric==3.2.2
GitPython==3.1.43
